backend/requirements.txt
flask>=3.0.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.3.0
reportlab>=4.1.0
gunicorn>=21.2.0
Flask-SQLAlchemy>=3.1.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
firebase-admin>=6.5.0
flask-cors>=4.0.0
